requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
boto3==1.34.162
botocore==1.34.162
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipython==8.18.1
jedi==0.19.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
kubernetes==32.0.1
langchain==0.2.16
langchain-aws==0.1.17
langchain-chroma==0.1.2
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.147
markdown-it-py==3.0.0
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openpyxl==3.1.5
opentelemetry-api==1.32.0
opentelemetry-exporter-otlp-proto-common==1.32.0
opentelemetry-exporter-otlp-proto-grpc==1.32.0
opentelemetry-instrumentation==0.53b0
opentelemetry-instrumentation-asgi==0.53b0
opentelemetry-instrumentation-fastapi==0.53b0
opentelemetry-proto==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
opentelemetry-util-http==0.53b0
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pip==21.3.1
posthog==3.24.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9
s3transfer==0.10.4
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
setuptools==59.6.0
shellingham==1.5.4
six==1.17.0
sklearn==0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.3
tenacity==8.5.0
termcolor==2.4.0
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing_extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0